Fix race condition after manager terminates
Suppose the following happens:

- A process is preforked
- User terminates sunwalker-box, causing the manager to terminate
- The stemcell is reparented to the reaper
- The reaper receives SIGCHLD and collects wait status from the manager
- The reaper exits the loop
- The stemcell notices EOF on controlling fd and raises SIGKILL
- The reaper receives SIGCHLD and collects wait status from the stemcell
- The reaper keeps running the loop
- The reaper attempts another wait(2) and fails due to ECHILD

The reason we don't get this in practice is that the reaper typically
receives SIGUSR1 before receiving the second SIGCHLD, which causes it to
correctly exit the event loop.

The conditions under which this race can be reproduced is unclear, but
it happened rather often on Yuki's device.